refactor(settings): created separate packages for RTDB and CF
