+++ packages/flamelink/cf/settings/index.ts
@@ -1,0 +1,1 @@
+import '@flamelink/sdk-settings-cf'
