+++ packages/flamelink/cf/settings/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "flamelink/cf/settings",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "typings": "../../public.d.ts"
+}
