+++ packages/flamelink/rtdb/settings/index.ts
@@ -1,2 +1,1 @@
-import '@flamelink/sdk-settings-cf'
 import '@flamelink/sdk-settings-rtdb'
