+++ packages/schemas-rtdb/src/index.ts
@@ -285,7 +285,7 @@ export const factory: FlamelinkFactory = context => {
   return api
 }
 
-const register: App.SetupModule = (context: App.Context) => {
+export const register: App.SetupModule = (context: App.Context) => {
   if (context.dbType === 'rtdb') {
     return factory(context)
   }
