+++ packages/settings-cf/README.md
@@ -1,0 +1,5 @@
+# @flamelink/sdk-settings-cf
+
+This is the Cloud Firestore settings module for the Flamelink JavaScript SDK.
+
+> This package should only be used via the official [flamelink](https://www.npmjs.com/package/flamelink) package.
