+++ packages/settings-cf/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@flamelink/sdk-settings",
+  "name": "@flamelink/sdk-settings-cf",
   "version": "1.0.0-alpha.22",
   "publishConfig": {
     "access": "public"
   },
-  "description": "Settings module for Flamelink JavaScript SDK",
+  "description": "Settings module for Flamelink JavaScript SDK - Cloud Firestore",
   "keywords": [
     "flamelink",
     "sdk",
@@ -26,9 +26,11 @@
   ],
   "dependencies": {
     "@flamelink/sdk-app": "^1.0.0-alpha.22",
-    "@flamelink/sdk-settings-cf": "^1.0.0-alpha.22",
-    "@flamelink/sdk-settings-rtdb": "^1.0.0-alpha.22",
-    "@flamelink/sdk-settings-types": "^1.0.0-alpha.22"
+    "@flamelink/sdk-app-types": "^1.0.0-alpha.22",
+    "@flamelink/sdk-settings-types": "^1.0.0-alpha.22",
+    "@flamelink/sdk-utils": "^1.0.0-alpha.21",
+    "compose-then": "^1.0.1",
+    "lodash": "^4.17.15"
   },
   "scripts": {
     "prebuild": "rimraf dist",
