+++ packages/settings-cf/src/vendor.d.ts
