+++ packages/settings-rtdb/README.md
@@ -1,0 +1,5 @@
+# @flamelink/sdk-settings-rtdb
+
+This is the Realtime Database settings module for the Flamelink JavaScript SDK.
+
+> This package should only be used via the official [flamelink](https://www.npmjs.com/package/flamelink) package.
