+++ packages/settings-rtdb/rollup.config.js
@@ -1,0 +1,4 @@
+import pkg from './package.json'
+import getRollupConfig from '../../tools/rollup-split-package-base'
+
+export default getRollupConfig(pkg)
