+++ packages/settings-rtdb/src/helpers.ts
